test: accounts wallet actions
