+++ frontend/src/tests/lib/utils/date.utils.spec.ts
@@ -38,14 +38,14 @@ describe("secondsToDuration", () => {
 
 describe("secondsToDate", () => {
   it("starts in 1970", () => {
-    expect(secondsToDate(0)).toBe("January 1, 1970");
+    expect(secondsToDate(0)).toBe("Jan 1, 1970");
   });
 
   it("returns day in number, month and year", () => {
     // We only support english for now
     const march25of2022InSeconds = Math.round(1648200639061 / 1000);
     const expectedDateText = secondsToDate(march25of2022InSeconds);
-    expect(expectedDateText).toContain("March");
+    expect(expectedDateText).toContain("Mar");
     expect(expectedDateText).toContain("2022");
     expect(expectedDateText).toContain("25");
   });
@@ -53,14 +53,14 @@ describe("secondsToDate", () => {
 
 describe("secondsToDateTime", () => {
   it("should return formatted start date and time in 1970", () => {
-    expect(secondsToDateTime(BigInt(0))).toEqual("January 1, 1970 12:00 AM");
+    expect(secondsToDateTime(BigInt(0))).toEqual("Jan 1, 1970 12:00 AM");
   });
 
   it("should return formatted date and time", () => {
     // We only support english for now
     const march25of2022InSeconds = Math.round(1648200639061 / 1000);
     const expectedDateText = secondsToDateTime(BigInt(march25of2022InSeconds));
-    expect(expectedDateText).toEqual("March 25, 2022 9:30 AM");
+    expect(expectedDateText).toEqual("Mar 25, 2022 9:30 AM");
   });
 });
